trader_bot/requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pyotp>=2.9.0
schedule>=1.2.1
Wealthsimple-API-Python>=0.1.0
